+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project uses the changelog in accordance with [keepchangelog](http://keepachangelog.com/). Please use this to write notable changes, which is not the same as git commit log...
 
 ## [unreleased][unreleased]
+ - Changed `hf legic *` - now uses NG instead (@iceman1001)
+ - Added `hf legic view` - view contents of LEGIC Prime dump files (@iceman1001)
  - Changed `hf mfu restore` - now takes bin/json as dump files (@iceman1001)
  - Added `hf mfu view` - view contents of MFU dump files (@iceman1001)
  - Changed `hf_mf_uidbruteforce` - added support for S70, enhance UID length management (@cactuschibre)